pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tts-v2"
version = "2.0.0"
description = "Production BFSI TTS pipeline — hexagonal (ports & adapters) architecture"
requires-python = ">=3.11"
dependencies = [
    "coqui-tts>=0.27.5",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
    "num2words>=0.5.13",
    "transformers>=4.35.0",
    "tokenizers>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ipykernel>=6.0.0",
    "ipywidgets>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
